feat: add shell script for test silo's methods
